FIX Moved duplicate similarity func to parent class
